[WIP] Added a whitelist for restricting the StatsD metrics
A whitelist and metric map for the metrics that are
sent by Storm / Threshold Engine to the Monasca
StatsD agent/daemon.

=======

/etc/monasca/thresh-config.yml

```
statsdConfig:
  host: localhost
  port: 8125
  debugmetrics: true
  dimensions: !!map
    service : monitoring
    component : storm
  metricmap: !!map
    someMetricName : someNewMetricName
  whitelist: !!seq
    - someMetricName
```

host: IP or host where the Monasca Agent running a StatsD is running that will consume
      the metrics produced by Storm / Threshold Engine

port: UDP port number where the Monasca Agent running a StatsD daemon that will consume
      the metrics produced by Storm / Threshold Engine

prefix: This string will be prefixed to the internal metric

dimensions: A map of key/value pairs that will be passed along as dimensions for each metric

whitelist: A list of metrics in the native name that Storm presents

metricmap: A mapping from the native Storm metric name to a user defined name.  The user
           defined name is what will appear in the Monasca data store.  If there is no
           mapping present and it is listed in the whitelist then it will be published
           with the native name.

Change-Id: I7bcefd03d02714ac42efd9b2d9cadb77907fa17e