Update patch set 5
Patch Set 5:

I've done my testing in our Vagrant environment with both Vertica and InfluxDB and everything appears to work. Will +1 after Deklan's comments are addressed and the SQL changes in ansible-monasca-schema are made. My assumption is that we can get this change merged in the next few days.

Patch-set: 5
Label: Verified=0