Evaluate SubAlarms immediately if possible
Many SubAlarm expressions can be evaluated immediately. If the
expression is MAX(m) > 10, a single value of m > 10 will cause the
SubAlarm to transition to the ALARM state regardless of any other
value of m that is received. Same is true of MIN. COUNT can be
evaluated immediately if the operator is > or >=. AVG and SUM
can't be evaluated until the end of the evaluation window since
the value could go up or down depending on the values received.

Also see if the sliding window for a SubAlarm can be slid when a
metric is received for the SubAlarm. This could allow the SubAlarm
to be evaluated faster than waiting for the tick tuple since that
is only received every 60 seconds.

Add unit tests for immediate SubAlarm evaluation.

Add unit tests for previously untested parts of SubAlarmStats

Change-Id: I989a82328fa4ccc04b49d203f70a1adc9fa4d3bb