Update patch set 6
Patch Set 6:

Hi Tomasz, Can we hold off until next week to merge this? The Mitaka release is closing and I need to apply a tag. I would prefer not to introduce this change in the official Mitaka release, as it is new functionality that is kinda coming in late.

Also, We need to get Craig Bryant to review this code as he has he most knowledge in this part of the pipeline.

Patch-set: 6
Label: Verified=0