Update patch set 15
Patch Set 15:

(11 comments)

I’ve finally had time to really look at this change and I have some comments that I think will simplify the implementation.

First, I need to articulate the use case to ensure the changes will match what we expect:

The operator wants to get an Alarm notification when  the number of warning messages in a given log file exceeds 2 per every 10 minute interval.

Implementation:
 1. A process is configured to watch the log file and send a measurement every time it sees a Warning message. This is a sporadic measurement because it is only sent when a Warning messages I seen.
 2. This Alarm Definition is created: count(log.warn_message, 600) > 2. I.e., we are receiving more than 2 warning messages per 10 minutes.

The expected behavior of the Threshold Engine:
 1. When every measurement is received, Thresh will increment count and if it is > 2, then it will change the Alarm state to ALARM
 2. At the end of very 10 minute interval, if the count is < 2 and the current Alarm state is ALARM, Thresh will change the state to OK
 3. The initial state of the Alarm will be OK
     It will never be UNDETERMINED any other time so OK is more consistent, however due to the way Thresh works, the first Measurement that creates the Alarm is actually lost so it can’t be used to determine the state. Would be nice to resolve that issue

Personally, I would like to separate out sporadic from long period metrics. Long period metrics can be handled by configuring Alarm Definitions to have long periods, so I believe we really only need to deal with sporadic metrics.

I suggest making these changes to the current implementation:
 1. Delete period and just add a sporadic property with a default of false but can be set to true
 2. Add the sporadic property to the MetricDefinition class, not the Metric class. This simplifies the Threshold Engine changes and supports the change below. The split of MetricDefinition properties vs Metric properties is not reflected in the actual measurement json received so it will be easy to make that change

Another change that would nice would be to save the value of sporadic in the MetricDefinition in the Alarm database. Then change the API to retrieve it and the CLI to display it. The advantage of that would be that the user can look at an alarm and see that it is being evaluated sporadically. It won’t make it completely obvious because it will be a property of the metrics associated with the alarm and not a property of the alarm itself, but at least there is a way for the user to tell the alarm is evaluated sporadically. Of course, if this change is made, the sporadic property of the sub alarm no longer needs to be stored in the database because it can be computed from the associated metrics.

Patch-set: 15
Label: Verified=0