Update patch set 15
Patch Set 15:

Hi Tomasz, We had a discussion here at HPE and I think we've come up with a way for handling all the use cases we've been discussing so far.

1. A sporadic metric, such as generated from the logging service. In this case, we would add a new field to the sub-alarm expression to specify that there isn't a period associated with the sub-alarm for which it would be allowed to transition to the UNDETERMINED state if metrics weren't received.

2. An alarm that you don't necessarily want to go back to the OK state automatically and must manually be set back to OK, after it transitions to ALARM. This would be a new attribute in the alarm definition. This handles the case that Witek just discussed. We are tentatively calling this a "latched" alarm, but the name is not necessarily the best. If you have a periodic metric, and it transitions to the ALARM state, it would stay/latch in that state until manually cleared using the API. Similarly, if you had a sporadic metric that transitioned, if the alarms transitions to the ALARM state, it would not be cleared if a new metric was sent that would normally have automatically have set it back again.

I think adding the "period" as an attribute of the metric was a reasonable direction, but it had some issues. I think this is something that we would potentially want to add to make it more clear to the user what the period of a metric is or whether it is a sporadic metric. However, it isn't required to do the alarm evaluation.

The main goal of the sporadic metric, which is a parameter of an alarm sub-expression, is to prevent the alarm from transitioning to the UNDETERMINED state.

The main goal of the "latched" attribute is to keep the alarm in the ALARM'd state. In other words, don't transition automatically.

When we discussed this today, we thought we needed an extra parameter on the alarm expression to specify if it is sporadic. Someone will need to remind me again why a value of -1 can't be used on the period.

Patch-set: 15
Label: Verified=0