Update patch set 15
Patch Set 15:

I just went back an re-read Witek's email. So, I think as he discusses there, we want to introduce a new attribute on the alarm definition to specify that it is latched/immutable. Once it is in the ALARM'd state it will stay there until it is manually set to OK.

Then, instead of specifying "sporadic" as part of the metric, we want to do this in the alarm sub-expression, using either a value of -1 or a new parameter to the alarm sub-expression in addition to the period. Something like 

max(logging.num_errors{}, sporadic=True) > 0

or

max(logging.num_errors{}, -1) > 0

Someone will just need to remind me why -1 wasn't acceptable.

We've already set the precedent that the period of an alarm sub-expression needs to be aligned with the period of a metric. For example, if a metric is being sent every 10 minutes, then the period of the alarm should be set to something like 20 minutes. Therefore, if it is a sporadic metric, creating an alarm sub-expression that matches the sporadic metric, also seems reasonable.

Patch-set: 15