Update patch set 15
Patch Set 15:

From Roland: Someone will just need to remind me why -1 wasn't acceptable.

The problem with -1 is that it doesn't handle my use case of "the number of warning messages in a given log file exceeds 2 per every 10 minute interval." That one has sporadic metrics with a period. If we define period =-1 to be sporadic, then we will never be able to support sporadic with a period.

From Witold:
"2. latched alarm
 max(logging.num_errors{}, -1) > 0"
 No, Roland suggests that a latched alarm is defined by a new attribute on the Alarm Definition. This makes sense because it is really the property of an Alarm, not of the metric associated with one of the sub alarms. It does, however, require a schema change

Patch-set: 15