Update patch set 15
Patch Set 15:

So, in summary, I think this is what we believe needs to be done.

1. Add new "latched" attribute to the alarm definition. If latched, the alarm will stay in the ALARM'd state until manually set back to OK.

2. Add "sporadic" attribute to an alarm expression. This will prevent the alarm sub-expression going to the UNDETERMINED state when no metrics are being sent. Possibly, this should be called "ignore_undetermined", as it is more descriptive. Actually, I prefer "ignore_undetermined" so I'll use it here.

So, in the logging use case from Witek, latched=True and ignore_undetermined=true.

In Craig's use case, latched=False and ignore_undetermined=true, and the period specifies when the alarm would transition back to OK if no metrics are still being sent. This allows the alarm sub-expression to return to the OK state if metrics are no longer being received after the period has been passed. This allows an alarm to automatically go back to OK without manual intervention.

Patch-set: 15