Update patch set 15
Patch Set 15:

Hi Tomasz, That is correct. Supplying the period with the metric is still useful to have. For example, if you query the metrics in the system, you would be able to tell from the metric definition it's period and whether it is sporadic. However, the information isn't used by the threshold engine.

Patch-set: 15