Update patch set 16
Patch Set 16:

(5 comments)

I've left some comments in the code. Other than possibly a name change, it all looks mostly right to me, but I'm not the expert on this code and Craig should weigh in if possible.

I'm not sure you are planning on adding the "latched" attribute in this review. It would be done in a separate review. Before you do, let's come up with a better name at least. Latched seems a little awkward and I'm assuming there is a better name. Remind me of "There are only two hard things in Computer Science: cache invalidation and naming things." -- Phil Karlton

In terms of implementation, "latched" would be defined in the alarm definition, not the alarm sub-expression. Basically, it would be a global on the alarm definition.

The alarm sub-expression, basically the code that you've modified so far, would I think remain the same. The sub-alarms would be calculated in the same way.

As a consequence, the place where "latched" would be implemented would I believe be in the AlarmThresholdingBolt, https://github.com/openstack/monasca-thresh/blob/master/thresh/src/main/java/monasca/thresh/infrastructure/thresholding/AlarmThresholdingBolt.java

Patch-set: 16
Label: Verified=0