Update patch set 2
Patch Set 2:

This looks fine to me. 

Note for other reviewers, the Apache Storm 1.0.0 release came out in a couple of weeks ago. See, http://storm.apache.org/2016/04/12/storm100-released.html.

The main reason for this change is that there is a bug in Storm that we havne't been able to resolve. We are hoping this release resolves it. Additionally, the 1.0.0 release has a number of significant improvements, such as performance.

We are hoping that this is OK with everyone and no one objects to upgrading Storm.

Craig, Doesn't DevStack need to be update with the new version too?

Patch-set: 2
Label: Workflow=0