Update patch set 16
Patch Set 16:

(3 comments)

Hi, I wrote back to some of your comments basically spoilering next patch set :). 

Anyway I think I've managed to ensure that if sporadic is the case for SubAlarm it will be put in AlarmState.OK immediately when created. I just hope that this assumption is correct.

The latched attribute is IMHO different piece of functionality. It has nothing to do with processing of sporadic alarms, so I'd like to keep it outside of this change.

Patch-set: 16