Update patch set 6
Patch Set 6:

Hi Witold, your example creates the alarm definition AFTER the metric was sent. If you went through this sequence with the old code:

 monasca alarm-definition-create test "max(log.critical{},deterministic) > 0"
 monasca metric-create log.critical 1.0
 monasca alarm-list

You would see an alarm be created, but it would have state OK because the measurement that created the Alarm was discarded because the measurement arrived at the MetricAggregationBolt before the SubAlarm creation event.

With the new code, that same sequence will show the Alarm in ALARM state because the measurement was saved and then added to the SubAlarm when the SubAlarm create event arrived.

We can't save all possible metrics in case they may be needed soon for a new Alarm Definition. That would overload the Threshold Engine.

Patch-set: 6