Implement the Last Function
The Alarm state is driven by the last measurement with the newest
timestamp. Use the value even if the measurement is older than the
oldest bucket.

Never evaluate Last except on receiving of a measurement.

Add tests to ensure this works.

The change is dependent on the monasca-common change and the
change to monasca-api to add the state field to sub_alarm.

Change-Id: Ib5123ed035018757a50d9ebeb7335fbca48054f2
Implements: Blueprint last-value