Update patch set 3
Patch Set 3:

the funny thing is: due to this fix, the zombie alarms on the alarm-state-transitions topic no longer occur. I have to check the logs whether there are still hints pointing to zombie alarms. And as you say: we yet have to understand how the original issue was introduced.

Patch-set: 3
Label: Verified=0