Update patch set 3
Patch Set 3:

That is caused by an alarm for which no alarm-definition could be found (since it is cleaned up now). So it seems like what is still missing is the removal of the alarm itself.

Looking at MetricAggregationBolt line 126 resp. 136, I suspect that those streams have been mixed up.

Patch-set: 3