Update patch set 5
Patch Set 5:

(1 comment)

The unit test would be ensuring that evaluateThreshold() did not throw an exception in this case. That could be done either by making it protected so it can be called from AlarmThresholdingBoltTest or by changing the catch() block in execute() to call a protected method with the Exception. The default protected method would do nothing, but AlarmThresholdingBoltTest would override that method so it could access the Exception and ensure nothing was thrown. The second is what I would suggest.

Patch-set: 5