Update patch set 6
Patch Set 6:

Anytime the "mvn package" is run a new file will be generated with third-party libs info. Normally it is written to the proj. build directory but I have changed that to the root directory. This way I can reference it from the README.md 
Might seem like a mess.. Maybe I should leave it in the build directory and create a separate markdown file in the root of the project that lists the third party libs? This however will not reflect if any third-party libs were added or removed.

Patch-set: 6
Reviewer: Gerrit User 27836 <27836@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0