Update patch set 6
Patch Set 6:

I would keep the configuration simple as it was in patchset 3. In the README I would just leave the note that the list of all third-party licenses is generated every time the package is generated and can be found in `target/generated-sources/license` directory. I would not commit the list.

Patch-set: 6
Label: Verified=0