Update patch set 2
Patch Set 2:

> This change should be tested manually. DevStack installs old
 > published package.

That's actually not correct. It's compiled from current source in DevStack, only the dependency `monasca-common` is downloaded from the published version.
It means the tempest tests job should give correct information about the component. I have updated the Zuul test configuration in [1] and rebased this change on it.

[1] https://review.openstack.org/634310

Patch-set: 2
Label: Verified=0