Update patch set 1
Patch Set 1: Code-Review+1

So for example, if we had very slow workers and a lot of metrics coming in, topology.max.spout.pending would have a significant effect. In this case the metrics would just sit in Kafka, rather than getting queued up in a pending state within Storm? The downside is that latency increases a bit, but probably not enough for us to care?

The change looks fine, but haven't had a chance to test yet. Thanks for looking at this.

Patch-set: 1
Label: Code-Review=+1